HTTP MockChain Part 1 (#194)
This commit introduces an http server that provides endpoints
that run actions in the WalletAPI class against wallets stored
in memory. The transactions are currently thrown away but in
part 2 we will apply/validate those transactions against a
mockchain stored in memory. We will also provide a control API
for the mockchain.

During this commit the binary serialization of UTXO types has
been changed to derived versions using DeriveGeneric, this was
done to make serialization and deserialization easier as it is
used in JSON marshalling, it also cleans up the code a bit.