HTTP MockChain Part 2
This commit adds new endpoints to trigger validation
and notifications.

The next stage is to create a client and try to fit
it into the Trace api. The only issue I see is assertions.
Since these are Haskell functions, we cannot serialize
them. It may be that we can remove them from the Trace API
and run them separately. We may need more endpoints for
this in order to see more state within the emulator.