Plutus IR: initial design draft
This is a first draft of a design document for Plutus IR (name
bikeshedding welcome in general).

We should bear in mind that we can change the details of Plutus IR much
more easily, so even if we're unsure we can start with one approach and
see if it works out.