language-plutus-core/plutus-ir: add default Pretty instances
While the full-on pretty-printing machinery is useful, and we do want it
for PIR (so we can e.g. get debug printing of uniques), it's nice to
have a default implementation of plain `Pretty` for "boring" usage.

As a side benefit, this lets me define a printing function that works
for both PLC and PIR, which will be helpful for my demo. (Possibly we
can do this already but I am unable to figure out how to.)