replace hint with runghc
hint took a long time to compile new code and for the server to
warm up, see https://github.com/haskell-hint/hint/issues/56
Additionally, only one request could be processed at a time.
By spawning a runghc process for every request we get around
both of these problems. The trade off is that there is a startup
overhead, however some rough timing has shown this to be negligeable.

One unfortunate side effect is that mkFunction needed to be replaced
by mkFunctions so to help the user with this change we print an
informative error message if they try to use the deprecated mkFunction.