plutus-tx: represent newtypes as their underyling type
Fixes #967.

Turns out my slightly awkward support for aliases in PIR is very useful
here.

The code for `Lift` is pretty awkward, but that module is horribly
awkward to begin with.

The motivation for this approach is included in a note, but I've
realised that we really need to do something like this, because GHC is
likely to start throwing away the content of coercions in the
not-too-distant future, which would thorougly break our old approach.

As a bonus, this is more efficient.