Awful hacks to make Haddock and doctests work
Haddock and doctests also run in interpreted mode, so need
`-fno-ignore-interface-pragmas` and `-fobject-code`. However, one cannot
simply set those options, of course.

Haddock: can't figure out a way to get Haddock to actuall use the
options, except setting them in the cabal file, in which case they will
apply to the real build too. Compromise: set `defer-errors`, which is
safe for both the real build and will not break the Haddock, since it
doesn't *run* the program. We gate this behind a flag and set it only in
CI.

Doctest: doesnt' like `-fobject-code`, we have to reimport the main
module and do some wiggling to get things working.