plutus-tx: support typeclasses and add some basic classes
Fixes #1109.

It turns out this works just fine. I'm sorry for putting you all through
the pain of not having typeclasses!

We can't use the classes from `base`, unsurprisingly, due to lack
of unfoldings for the instance definitions.

However, we can define our own replacements, which is a bit of
boilerplate but not too bad.

Delightfully, you seem to generally be able to get away with
newtype-deriving them:
- We *do* handle newtype coercions, so the generated instance
definitions are actually fine.
- They seem to be small enough that GHC always puts in an unfolding for
them even though we can't annotate them.
    - This is a little dicey, but seems to be okay so far.

The class operations themselves annoyingly don't have unfoldings, but we
can recreate the RHSs using GHC's functions. Antother annoyance is #978:
this bites us for e.g. the default method declarations in `Eq`, which
means that `(==)` and `(/=)` are mutually recursive! So you need to put
your `Eq` instances in a separate file from the final use site.

To demonstrate, I added `Eq`, `Ord`, and `Functor` and plumbed them
through. I think the result is a lot better!

We get a ~10% increase in script size and a comparable slowdown. I think
this is somewhat inevitable since we're not taking much advantage of
GHC's specializer, so we are paying a price from passing all the
dictionaries around.

TODO (but probably not in this PR):
- `Semigroup`/`Monoid`
- `Applicative`/`Monad`
    - With `RebindableSyntax` I think we could even do `do` notation!
- `Num`?
    - I'm unsure whether we want this, given that `Num` is kind of a
      dodgy class, but I would like to be able to use `+` and `-`...
    - Possibly we should define our own `Num`-like class.
- More tuple instances
    - I was lazy and only went to 2-tuples.