plutus-tx: reverse GHC's OccName tidying
I've observed these being nondeterministic on bigger examples (possibly
related to the order in which modules are processed). So we kind of have
to strip them.

This has two annoying downsides:
- We strip "real" trailing digits from names (witness the diffs in the
  `plutus-tx` tests).
- We lose the disambiguation they provide. This is especially annoying
  for generated names like the many occurrences of `ds`.

We could regain the disambiguation by running our *own* disambiguation
pass, which would be deterministic. Not sure if this is worth it, I'll
probably just see how annoying it is in practice.