runghc was creating *.hi and *.o files
These files were not being deleted hence they filled up the disk
eventually. We now delete the whole directory.

Additionally the temporary library has been updated and is now
generalized to MonadIO however we needed to use withFile from
System.IO and this did need to be generalized to MonadIO