plutus-emulator: add a direct Failure event
This is useful for putting in programmatic assertiosn that don't rely on
access to runtime stuff we don't know yet.

For example, we can now write a function that asserts that a wallet
action was sucessful, so we can easily get the value returned from it.
This is handy in a few places where wallet actions compute values and we
would otherwise have to recompute them, for example the test I changed
in this commit.