ledger: move application of 'check' into 'wrapValidator'
This finally makes the type of validators on-chain independent of our
datatype encoding scheme, since they can now return anything and must
signal failure via an error.

I haven't changed the type in the EUTXO spec, because that specifies the
type of the *interface* to validation, which can wrap up the error
handling and actually return a bool.

As a side effect, this removes the only use of the Plutus Tx plugin in
`plutus-wallet-api` (see https://github.com/input-output-hk/plutus/pull/1557
for why this is desirable).