ledger: move full data value to transaction outputs
Previously, data values were attached to transaction outputs in
full. This is bad for the UTXO size. Instead we're moving to having
just the *hash* of the data value on the output, and having the
transaction *input* provide the full data.

However, in order to make it feasible for people to find out what the
data values are, we're also allowing them to be recorded on the
transaction that creates an output which uses that data value. So
wallets can keep a richer index (for us, `AddressMap` does this work)
which keeps the transaction as well as the transaction output, and so
can look up the corresponding data value.

This necessitates a lot of awkward plumbing. It also adds a bunch of
landmines for people trying to directly construct 'Tx's: you can get the
hashes wrong, the data vaulues wrong, or forget to attach the evidence
to the creating transaction. We should really try and get people to use
the higher-level functions that handle this for you.