plutus-ir: share fixpoint combinator
Fixes #1657.

These can be quite big, and otherwise we were making one per recursive
function. We still have one per arity of recursive function, but it
seems to save us ~10% on program size, which is nice.

The `language-plutus-core` changes are pretty minimal: `getMutualFixOf`
lets us provide the fixpoint function ourselves (or - a variable
reference to one), and `fixN` also provides the type, which we need for
creating variable bindings.

I think we want to use `DefT` some more in the PIR compilation pipeline,
particularly for sharing all the units in `NonStrict`, but I've kept it
simple for now.