plutus-core: remove the value restriction in favour of big-lambdas computing
This removes the value restriction from Plutus Core. This requires us to
decide when the bodies of type abstractions reduce. We choose to say that
they work like the bodies of lambda abstractions: they reduce after the
substitution created by the abstraction is applied.

As it happens, this requires few changes:
- The reduction rules for type abstractions already reveal the body
after the substitution and not before - since there is no reduction
frame for type abstractions, this has the behaviour we want.
- The specification of the CK machine is changed to compute the body of
the abstraction after the instantiation.
    - As it turns out, the implementations of all our machines
already do this. Currently this is redundant since it will already be a
value, but it means we don't need to change anything.
- We don't need to say anything about erasure, since we have neither a
specification of it nor an implementation at the moment. There will need
to be change in the metatheory, however.