PIR: add concept of a pure term, use instead of isTermValue
For PIR, we really care about whether a term can have effects or not,
since this determines whether we can move or remove it. So we talk about
purity instead of whether the term is a "value" to avoid confusion.

Pure terms include variables and stuck builtin applications, which are cases
we care about. Handling the stuck builtins involves a lot of extra plumbing
to get the builtin info to where we need it, but this is okay as we'll
need to do some of this plumbing anyway for the typechecker.

Variables are also a little tricky, since PIR has non-strict variables
that can have effects at their use site. We have to be conservative in
many cases, but we can sometimes do better, e.g. the dependency analysis
pass now tracks the strictness of variables as it goes, so it can be
more precise.