quickfix to marlowe editor for simplifying values, remove parseContract call
This was a bit tricky, we need access to the lint result in the code
actions provider however we don't have access to the contract at this
time, so we can't even do another parse and lint. The issue though is that the
actions provider is closed over by the monaco javascript code at
initialisation. We therefore need to provide an exposed reference to the
actions provider. Additionally I wanted to keep any marlowe specific
code out of the generic monaco code. What I chose to do was to keep a
reference to the code provider in the Halogen state. This way you can
provide a custom implementation of the CodeActionProvider that injects
state into it dynamically. In our case we inject the parsed contract and
the warnings in whenever the markers are generated (since this is done
outside of monaco). Now when the actions provider is called it has
access to the latest parsed contract and warnings so that it can build
code actions.

This has an additional benefit of getting rid of a call to parseContract
that is made an extra time every time we type. There is still one more
place where parseContract is called, that is updateContractInStateP. I
think in the future we can get rid of this as well by using the same
technique.