Fixing a bug with the Data.String.Extra.abbreviate tests.
Turns out it was the test at fault, in two ways. It was supposed to
compare applying the function once, and applying it twice, and expect
the same result. In reality, was comparing double-abbreviation with no
abbrevation at all, but choosing such a large maximum string length
that it didn't matter, most of the time. :laugh: :sob: