Refactor for restricted mode
Apply some changes to allow evaluation in restricted mode. More
specifically this changes the npmlock2nix.node_modules invocation such
that local paths for `package.json` and `package-lock.json` can be
specified. This is necessary because in restricted mode it isn't
possible to readFile something from a nix store path.